chore: combine migrations, fix seeding and testing (#15)
The migration files had logic in them that really should not be there,
so they are now consolidated and logic is removed. This will require
dumping and restoring the DB when pushed to prod.

The seeders now only seed a few users and the games for the site. List
seeding doesn't work too well since it needs files. Factories still
exist for the purposes of automated testing, which has also been fixed.